boardRepository 테스트를 추출
