+++ src/main/java/uhs/alphabet/board/BoardRepository.java
@@ -4,8 +4,8 @@ import uhs.alphabet.board.BoardEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
+import java.util.Optional;
 
 public interface BoardRepository extends JpaRepository<BoardEntity, Long> {
     List<BoardEntity> findByTitleContaining(String key);
-
 }
